Add make targets cluster-sync-cdi & cluster-sync-test-infra

+++ cluster-sync/sync.sh
@@ -111,6 +111,20 @@ function wait_cdi_pods_updated {
   fi
 }
 
+# Start functional test HTTP server.
+# We skip the functional test additions for external provider for now, as they're specific
+if [ "${KUBEVIRT_PROVIDER}" != "external" ] && [ "${CDI_SYNC}" == "test-infra" ]; then
+  _kubectl delete all -n cdi -l cdi.kubevirt.io/testing
+  _kubectl apply -f "./_out/manifests/bad-webserver.yaml"
+  _kubectl apply -f "./_out/manifests/file-host.yaml"
+  _kubectl apply -f "./_out/manifests/registry-host.yaml"
+  # Imageio test service:
+  _kubectl apply -f "./_out/manifests/imageio.yaml"
+  # vCenter (VDDK) test service:
+  _kubectl apply -f "./_out/manifests/vcenter.yaml"
+  exit 0
+fi
+
 mkdir -p ./_out/tests
 rm -f $OLD_CDI_VER_PODS $NEW_CDI_VER_PODS
 
@@ -197,15 +211,3 @@ crds+=($operator_crds)
 check_structural_schema "${crds[@]}"
 
 configure_storage
-
-# Start functional test HTTP server.
-# We skip the functional test additions for external provider for now, as they're specific
-if [ "${KUBEVIRT_PROVIDER}" != "external" ]; then
-  _kubectl apply -f "./_out/manifests/bad-webserver.yaml"
-  _kubectl apply -f "./_out/manifests/file-host.yaml"
-  _kubectl apply -f "./_out/manifests/registry-host.yaml"
-  # Imageio test service:
-  _kubectl apply -f "./_out/manifests/imageio.yaml"
-  # vCenter (VDDK) test service:
-  _kubectl apply -f "./_out/manifests/vcenter.yaml"
-fi
